Use translated place names on bio pages
